pluginhandler: support more complex stage-packages. (#1059)
Currently snapcraft only supports a flat list of stage packages to be
staged regardless of target architecture. This commit introduces a more
complex grammar that allows one to filter stage packages depending on
various selectors (target arch for right now), as well as specify
optional packages.

The grammar is made up of two statements: `on` and `try`.

- on <selector>[,<selector>...]
  - ...
- else[ fail]:
  - ...

The body of the `on` clause is taken into account if every (AND, not OR)
selector is true for the target build environment. Currently the only
selectors supported are target architectures (e.g. amd64).

If the `on` clause doesn't match and it's immediately followed by an
`else` clause, the `else` clause must be satisfied. An `on` clause
without an `else` clause is considered satisfied even if no selector
matched. The `else fail` form allows erroring out if an `on` clause was
not matched.

- try:
  - ...
- else:
  - ...

The body of the `try` clause is taken into account only when all
packages contained within it are valid. If not, if it's immediately
followed by `else` clauses they are tried in order, and one of them must
be satisfied. A `try` clause with no `else` clause is considered
satisfied even if it contains invalid packages.

LP: #1637282

Signed-off-by: Kyle Fazzari <kyle@canonical.com>